third_party/blink/web_tests/clipboard/async-clipboard/async-write-blobtext-read-text.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard write ([text/plain Blob]) -> readText tests</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../http/tests/resources/permissions-helper.js"></script>
<script>
async function readWriteTest(textInput) {
  promise_test(async t => {
    await PermissionsHelper.setPermission('clipboard-read', 'granted');
    await PermissionsHelper.setPermission('clipboard-write', 'granted');

    const blobInput = new Blob([textInput], {type: 'text/plain'});

    await navigator.clipboard.write({'text/plain': blobInput});
    const textOutput = await navigator.clipboard.readText();

    assert_equals(textOutput, textInput);
  }, 'Verify write and read clipboard given text: ' + textInput);
}

readWriteTest('Clipboard write ([text/plain Blob]) -> read text test');
readWriteTest('non-Latin1 text encoding test データ');
</script>
